requirements.txt
chainlit==0.7.700
groq
python-dotenv
transformers
torch
accelerate
